@@ -1,3 +1,3 @@
 module github.com/artyom/buffering
 
-go 1.12
+go 1.19
